Rename the plugin to GroupResolver.
