+++ src/GroupResolverManager.php
@@ -9,10 +9,10 @@ use Drupal\Core\Extension\ModuleHandlerInterface;
 /**
  * Provides the OG context plugin manager.
  */
-class OgContextManager extends DefaultPluginManager {
+class GroupResolverManager extends DefaultPluginManager {
 
   /**
-   * Constructor for OgContextManager objects.
+   * Constructor for GroupResolverManager objects.
    *
    * @param \Traversable $namespaces
    *   An object that implements \Traversable which contains the root paths
@@ -23,10 +23,10 @@ class OgContextManager extends DefaultPluginManager {
    *   The module handler to invoke the alter hook with.
    */
   public function __construct(\Traversable $namespaces, CacheBackendInterface $cache_backend, ModuleHandlerInterface $module_handler) {
-    parent::__construct('Plugin/OgContext', $namespaces, $module_handler, 'Drupal\og\OgContextInterface', 'Drupal\og\Annotation\OgContext');
+    parent::__construct('Plugin/GroupResolver', $namespaces, $module_handler, 'Drupal\og\OgContextInterface', 'Drupal\og\Annotation\GroupResolver');
 
-    $this->alterInfo('og_context_og_context_info');
-    $this->setCacheBackend($cache_backend, 'og_context_og_context_plugins');
+    $this->alterInfo('og_group_resolver_info');
+    $this->setCacheBackend($cache_backend, 'og_group_resolver_plugins');
   }
 
 }
