+++ tests/modules/og_test/src/Plugin/GroupResolver/UserGroupReference.php
@@ -1,10 +1,10 @@
 <?php
 
-namespace Drupal\og_test\Plugin\OgContext;
+namespace Drupal\og_test\Plugin\GroupResolver;
 
 /**
  * @file
- * Contains \Drupal\og_test\Plugin\OgContext\UserGroupReference.
+ * Contains \Drupal\og_test\Plugin\OgResolver\UserGroupReference.
  */
 
 use Drupal\og\OgContextBase;
@@ -12,7 +12,7 @@ use Drupal\og\OgContextBase;
 /**
  * Get the group from a field reference attached to the current user.
  *
- * @OgContext(
+ * @GroupResolver(
  *  id = "user_group_reference",
  *  label = "User group reference",
  *  description = @Translation("A dummy plugin to return the first group from a field attach to the user.")
